projects/logreg_playground/requirements.txt
numpy
scipy
pandas
pyyaml
pyarrow
matplotlib
scikit-learn
